// src/interfaces/GameEvent.ts
export interface MetaData {
	wait: number;
	timeStamp: string;
}
// ...
export interface GameData {
	game: Game;
	datetime: Datetime;
	status: Status;
	teams: Teams;
	players: Players;
	venue: Venue3;
}
// ...
export interface LiveData {
	plays: Plays;
	linescore: Linescore;
	boxscore: Boxscore;
	decisions: Decisions;
}

export interface GameEventsData {
	copyright: string;
	gamePk: number;
	link: string;
	metaData: MetaData;
	gameData: GameData;
	liveData: LiveData;
}

export default interface GameEvents {
	data: GameEventsData;
}